Module-1-Working_with_LangChain_and_LCEL/Legal_doc_review/requirements.txt

streamlit
PyPDF2
faiss-cpu
langchain[google-genai, faiss]
python-dotenv
